+++ lib/transpec/dynamic_analyzer/runtime_data.rb
@@ -1,18 +1,26 @@
 # coding: utf-8
 
+require 'json'
+require 'ostruct'
 require 'pathname'
 
 module Transpec
   class DynamicAnalyzer
     class RuntimeData
-      attr_reader :hash
+      attr_reader :data
 
-      def initialize(hash = {})
-        @hash = hash
+      def self.load(string_or_io)
+        options = { object_class: CompatibleOpenStruct, symbolize_names: true }
+        data = JSON.load(string_or_io, nil, options)
+        new(data)
+      end
+
+      def initialize(data = CompatibleOpenStruct.new)
+        @data = data
       end
 
       def [](node)
-        @hash[node_id(node)]
+        @data[node_id(node)]
       end
 
       def node_id(node)
@@ -21,6 +29,21 @@ module Transpec
         absolute_path = File.expand_path(source_buffer.name)
         relative_path = Pathname.new(absolute_path).relative_path_from(Pathname.pwd).to_s
         [relative_path, source_range.begin_pos, source_range.end_pos].join('_')
+      end
+
+      class CompatibleOpenStruct < OpenStruct
+        # OpenStruct#[] is available from Ruby 2.0.
+        unless method_defined?(:[])
+          def [](key)
+            __send__(key)
+          end
+        end
+
+        unless method_defined?(:[]=)
+          def []=(key, value)
+            __send__("#{key}=", value)
+          end
+        end
       end
     end
   end